refactor: 음식점 데이터 타입에 대한 네임을 구체적으로 변경
